Changed object id type from "long long" to "uint64_t".
git-svn-id: https://svn.code.sf.net/p/stemkit/code@952 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/rt/Thread.h
@@ -55,7 +55,7 @@ namespace sk {
 
         // sk::util::Object re-implementation.
         const sk::util::Class getClass() const;
-        long long getId() const;
+        uint64_t getId() const;
 
         static thread::Generic& currentThread();
         static int activeCount();
